+++ src/messages/types.ts
@@ -1,0 +1,7 @@
+import { Envelope } from '@cucumber/messages'
+import { ReadonlyDeep } from 'type-fest'
+
+export type EnvelopeFromFile = {
+  file: string
+  envelope: ReadonlyDeep<Envelope>
+}