+++ src/runner/MessagesCollector.ts
@@ -1,41 +1,19 @@
-import { TestContext } from 'node:test'
-
 import { Envelope } from '@cucumber/messages'
 
-export const PROTOCOL_PREFIX = '@cucumber/messages:'
+import { EnvelopesReplaySubject } from '../messages/index.js'
 
-export interface MessagesCollector {
-  connect(nodeTestContext: TestContext): void
-  push(...envelopes: ReadonlyArray<Envelope>): void
-}
-
-export class DiagnosticMessagesCollector implements MessagesCollector {
-  private queue: Array<Envelope> = []
-  private context: TestContext | undefined
-
-  private serialise(envelope: Envelope) {
-    return PROTOCOL_PREFIX + JSON.stringify(envelope)
-  }
-
-  connect(context: TestContext) {
-    this.context = context
-    if (this.queue.length) {
-      this.queue.splice(0).forEach((envelope) => context.diagnostic(this.serialise(envelope)))
-    }
-  }
+export class MessagesCollector {
+  constructor(
+    private readonly file: string,
+    private readonly subject: EnvelopesReplaySubject
+  ) {}
 
   push(...envelopes: ReadonlyArray<Envelope>) {
-    if (this.context) {
-      for (const envelope of envelopes) {
-        this.context.diagnostic(this.serialise(envelope))
-      }
-    } else {
-      this.queue.push(...envelopes)
-    }
+    envelopes.forEach((envelope) =>
+      this.subject.next({
+        file: this.file,
+        envelope,
+      })
+    )
   }
-}
-
-export class NoopMessagesCollector implements MessagesCollector {
-  connect() {}
-  push() {}
 }