+++ test/integration/reporters.spec.ts
@@ -1,9 +1,9 @@
-import { stripVTControlCharacters } from 'node:util'
-import { expect } from 'chai'
-import { makeTestHarness } from '../utils.js'
 import path from 'node:path'
-import { TestStepResultStatus } from '@cucumber/messages'
-import { Query } from '@cucumber/query'
+import { stripVTControlCharacters } from 'node:util'
+
+import { expect } from 'chai'
+
+import { makeTestHarness } from '../utils.js'
 
 describe('Reporters', () => {
   describe('spec', () => {
@@ -92,48 +92,6 @@ Given('a step', () => {})`
     t.todo();
   });
 `)
-    })
-  })
-
-  describe('message', () => {
-    it('only factors cucumber tests into results', async () => {
-      const harness = await makeTestHarness()
-      await harness.writeFile(
-        'features/first.feature',
-        `Feature:
-  Scenario:
-    Given a step
-`
-      )
-      await harness.writeFile(
-        'features/steps.js',
-        `import { Given } from '@cucumber/node'
-Given('a step', () => {})
-`
-      )
-      await harness.writeFile(
-        'example.test.mjs',
-        `import test from 'node:test'
-test('top level', (t) => {
-    test('next level', (t1) => {
-        t1.test('failing test', (t2) => {
-            t2.assert.strictEqual(1, 2)
-        })
-    })
-})
-`
-      )
-
-      const query = new Query()
-      await harness.run(query)
-
-      expect(query.findAllTestCaseStarted().length).to.eq(1)
-      expect(
-        query
-          .findAllTestCaseStarted()
-          .map((testCaseStarted) => query.findMostSevereTestStepResultBy(testCaseStarted)?.status)
-      ).to.deep.eq([TestStepResultStatus.PASSED])
-      expect(query.findTestRunFinished()?.success).to.be.true
     })
   })