+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Avoid diagnostic hack for messages ([#177](https://github.com/cucumber/cucumber-node/pull/177))
 
 ## [0.6.1] - 2025-12-14
 ### Fixed